Add files via upload
In this notebook, I'm going through an machine learning project with the goal of predicting the sale price of bulldozers.

Since I'm predicting a number, this kind of problem is known as a regression problem. The data and evaluation metric we'll be using (root mean square log error or RMSLE) is from the Kaggle Bluebook for Bulldozers competition.